@@ -1,7 +1,7 @@
 [package]
 name = "bevy-inspector-egui"
-version = "0.30.0"
-edition = "2021"
+version = "0.31.0"
+edition = "2024"
 repository = "https://github.com/jakobhellermann/bevy-inspector-egui/"
 readme = "README.md"
 license = "MIT OR Apache-2.0"
@@ -72,10 +72,12 @@ uuid = "1.11.0"
 bevy = { path = "../../../../bevy_engine", default-features = false, features = [
     "x11",
     "bevy_winit",
+    "bevy_state",
     "bevy_window",
     "bevy_pbr",
     "bevy_sprite",
     "bevy_core_pipeline",
+    "bevy_picking",
     "animation",
     "png",
     "zstd",
@@ -83,7 +85,8 @@ bevy = { path = "../../../../bevy_engine", default-features = false, features = 
     "ktx2",
 ] }
 egui_dock = "0.16"
-# transform-gizmo-egui = "0.1"
+bevy_math = { version = "0.16.0", features = ["mint"] }
+transform-gizmo-egui = "0.5"
 # bevy_mod_picking = { git = "https://github.com/aevyrie/bevy_mod_picking", rev = "554649a951689dce66d0d759839b326874e8826f", default-features = false, features = ["backend_raycast", "backend_egui", "backend_sprite"] }
 # bevy_framepace = "0.11"
 
@@ -131,6 +134,3 @@ path = "examples/integrations/egui_dock.rs"
 [[example]]
 name = "side_panel"
 path = "examples/integrations/side_panel.rs"
-
-[lints.rust]
-unexpected_cfgs = { level = "warn", check-cfg = ['cfg(egui_dock_gizmo)'] }
