@@ -3,7 +3,6 @@ use std::collections::HashSet;
 use crate::bevy_inspector::{EntityFilter, Filter};
 use crate::utils::guess_entity_name;
 use bevy_ecs::{prelude::*, query::QueryFilter};
-use bevy_hierarchy::{Children, Parent};
 use bevy_reflect::TypeRegistry;
 use egui::{CollapsingHeader, RichText};
 
@@ -62,14 +61,16 @@ impl<T> Hierarchy<'_, T> {
         QF: QueryFilter,
         F: EntityFilter,
     {
-        let mut root_query = self.world.query_filtered::<Entity, (Without<Parent>, QF)>();
+        let mut root_query = self
+            .world
+            .query_filtered::<Entity, (Without<ChildOf>, QF)>();
 
         let always_open: HashSet<Entity> = self
             .selected
             .iter()
             .flat_map(|selected| {
                 std::iter::successors(Some(selected), |&entity| {
-                    self.world.get::<Parent>(entity).map(|parent| parent.get())
+                    self.world.get::<ChildOf>(entity).map(|c| c.0)
                 })
                 .skip(1)
             })
